tools/depcheck/requirements.txt
requests
beautifulsoup4
packaging
tqdm
pyyaml
